+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `updateOrderFormMarketingData` on checkout mutations
+
 ## [2.136.6] - 2020-12-03
 
 ### Changed